Added lib file and moved  utility functions
