+++ package.json
@@ -11,7 +11,7 @@
   "packageManager": "npm@10.7.0",
   "overrides": {
     "react": "$react",
-    "react-dom": "19.0.0-rc-66855b96-20241106"
+    "react-dom": "$react-dom"
   },
   "dependencies": {
     "@sendgrid/mail": "^8.1.4",
